[Core] Extract payment target transition resolver
