+++ src/Sylius/Component/Core/StateResolver/OrderPaymentStateResolver.php
@@ -15,8 +15,8 @@ use SM\Factory\FactoryInterface;
 use SM\StateMachine\StateMachineInterface;
 use Sylius\Component\Order\Model\OrderInterface;
 use Sylius\Component\Order\StateResolver\StateResolverInterface;
-use Sylius\Component\Core\Model\PaymentInterface;
 use Sylius\Component\Core\OrderPaymentTransitions;
+use Sylius\Component\Order\StateResolver\TargetTransitionResolverInterface;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
@@ -31,11 +31,18 @@ class OrderPaymentStateResolver implements StateResolverInterface
     private $stateMachineFactory;
 
     /**
-     * @param FactoryInterface $stateMachineFactory
+     * @var TargetTransitionResolverInterface
      */
-    public function __construct(FactoryInterface $stateMachineFactory)
+    private $transitionResolver;
+
+    /**
+     * @param FactoryInterface $stateMachineFactory
+     * @param TargetTransitionResolverInterface $transitionResolver
+     */
+    public function __construct(FactoryInterface $stateMachineFactory, TargetTransitionResolverInterface $transitionResolver)
     {
         $this->stateMachineFactory = $stateMachineFactory;
+        $this->transitionResolver = $transitionResolver;
     }
 
     /**
@@ -44,7 +51,7 @@ class OrderPaymentStateResolver implements StateResolverInterface
     public function resolve(OrderInterface $order)
     {
         $stateMachine = $this->stateMachineFactory->get($order, OrderPaymentTransitions::GRAPH);
-        $targetTransition = $this->getTargetTransition($order);
+        $targetTransition = $this->transitionResolver->resolve($order);
 
         if (null !== $targetTransition) {
             $this->applyTransition($stateMachine, $targetTransition);
@@ -60,58 +67,5 @@ class OrderPaymentStateResolver implements StateResolverInterface
         if ($stateMachine->can($transition)) {
             $stateMachine->apply($transition);
         }
-    }
-
-    /**
-     * @param OrderInterface $order
-     *
-     * @return string|null
-     */
-    private function getTargetTransition(OrderInterface $order)
-    {
-        $refundedPaymentTotal = 0;
-        $refundedPayments = $this->getPaymentsWithState($order, PaymentInterface::STATE_REFUNDED);
-
-        foreach ($refundedPayments as $payment) {
-            $refundedPaymentTotal += $payment->getAmount();
-        }
-
-        if (0 < $refundedPayments->count() && $refundedPaymentTotal >= $order->getTotal()) {
-            return OrderPaymentTransitions::TRANSITION_REFUND;
-        }
-
-        if ($refundedPaymentTotal < $order->getTotal() && 0 < $refundedPaymentTotal) {
-            return OrderPaymentTransitions::TRANSITION_PARTIALLY_REFUND;
-        }
-
-        $completedPaymentTotal = 0;
-        $completedPayments = $this->getPaymentsWithState($order, PaymentInterface::STATE_COMPLETED);
-
-        foreach ($completedPayments as $payment) {
-            $completedPaymentTotal += $payment->getAmount();
-        }
-
-        if (0 < $completedPayments->count() && $completedPaymentTotal >= $order->getTotal()) {
-            return OrderPaymentTransitions::TRANSITION_PAY;
-        }
-
-        if ($completedPaymentTotal < $order->getTotal() && 0 < $completedPaymentTotal) {
-            return OrderPaymentTransitions::TRANSITION_PARTIALLY_PAY;
-        }
-
-        return null;
-    }
-
-    /**
-     * @param OrderInterface $order
-     * @param string $state
-     *
-     * @return PaymentInterface[]
-     */
-    private function getPaymentsWithState(OrderInterface $order, $state)
-    {
-        return $order->getPayments()->filter(function (PaymentInterface $payment) use ($state) {
-            return $state === $payment->getState();
-        });
     }
 }
