+++ src/Sylius/Component/Core/StateResolver/PaymentTargetTransitionResolver.php
@@ -1,0 +1,79 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sylius\Component\Core\StateResolver;
+
+use Sylius\Component\Order\Model\OrderInterface;
+use Sylius\Component\Core\Model\PaymentInterface;
+use Sylius\Component\Core\OrderPaymentTransitions;
+use Sylius\Component\Order\StateResolver\TargetTransitionResolverInterface;
+
+/**
+ * @author Paweł Jędrzejewski <pawel@sylius.org>
+ * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
+ * @author Grzegorz Sadowski <grzegorz.sadowski@lakion.com>
+ * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
+ */
+final class PaymentTargetTransitionResolver implements TargetTransitionResolverInterface
+{
+    /**
+     * @param OrderInterface $order
+     *
+     * @return string|null
+     */
+    public function resolve(OrderInterface $order)
+    {
+        $refundedPaymentTotal = 0;
+        $refundedPayments = $this->getPaymentsWithState($order, PaymentInterface::STATE_REFUNDED);
+
+        foreach ($refundedPayments as $payment) {
+            $refundedPaymentTotal += $payment->getAmount();
+        }
+
+        if (0 < $refundedPayments->count() && $refundedPaymentTotal >= $order->getTotal()) {
+            return OrderPaymentTransitions::TRANSITION_REFUND;
+        }
+
+        if ($refundedPaymentTotal < $order->getTotal() && 0 < $refundedPaymentTotal) {
+            return OrderPaymentTransitions::TRANSITION_PARTIALLY_REFUND;
+        }
+
+        $completedPaymentTotal = 0;
+        $completedPayments = $this->getPaymentsWithState($order, PaymentInterface::STATE_COMPLETED);
+
+        foreach ($completedPayments as $payment) {
+            $completedPaymentTotal += $payment->getAmount();
+        }
+
+        if (0 < $completedPayments->count() && $completedPaymentTotal >= $order->getTotal()) {
+            return OrderPaymentTransitions::TRANSITION_PAY;
+        }
+
+        if ($completedPaymentTotal < $order->getTotal() && 0 < $completedPaymentTotal) {
+            return OrderPaymentTransitions::TRANSITION_PARTIALLY_PAY;
+        }
+
+        return null;
+    }
+
+    /**
+     * @param OrderInterface $order
+     * @param string $state
+     *
+     * @return PaymentInterface[]
+     */
+    private function getPaymentsWithState(OrderInterface $order, $state)
+    {
+        return $order->getPayments()->filter(function (PaymentInterface $payment) use ($state) {
+            return $state === $payment->getState();
+        });
+    }
+}
